Refactor component folders
