+++ components/Wami/WordCard.js
